requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pandas
scikit-learn
python-dotenv

